since we need to model storage capacity as a function of width and length we had to change  "int permlanes"  to "double permlanes". however, the method BasicLink.getLanes() returns still "int", so no other classes should be affected ...
git-svn-id: file:///srv/local/data/cleanRepo/trunk@568 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/network/Link.java
@@ -55,7 +55,7 @@ public class Link extends BasicLink {
 		this.length = Double.parseDouble(length);
 		this.freespeed = Double.parseDouble(freespeed);
 		this.capacity = Double.parseDouble(capacity);
-		this.permlanes = Integer.parseInt(permlanes);
+		this.permlanes = Double.parseDouble(permlanes);
 		this.origid = origid;
 		this.type = type;
 		this.euklideanDist = ((Node)this.from).getCoord().calcDistance(((Node)this.to).getCoord());
